add changes

+++ database/migrations/2021_08_17_052620_create_posts_table.php
@@ -13,44 +13,48 @@ class CreatePostsTable extends Migration
      */
     public function up()
     {
-        Schema::create('posts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')
-                  ->constrained()
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
-            $table->string('title');
-            $table->string('slug')->unique();
-            $table->text('excerpt');
-            $table->text('body');
-            $table->boolean('featured')->default(false);
-            $table->string('image');
-            $table->timestamps();
-        });
-
-        Schema::create('category_post', function(Blueprint $table) {
-            $table->id();
-            $table->foreignId('category_id')
-                  ->constrained()
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
-            $table->foreignId('post_id')
-                  ->constrained()
-                  ->onUpdate('cascade')
-                  ->onDelete('cascade');
-        });
-
-        Schema::create('post_tag', function(Blueprint $table) {
-            $table->id();
-            $table->foreignId('post_id')
-                  ->constrained()
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
-            $table->foreignId('tag_id')
-                  ->constrained()
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
-        });
+        if (!Schema::hasTable('posts')) {
+            Schema::create('posts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')
+                    ->constrained()
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+                $table->string('title');
+                $table->string('slug')->unique();
+                $table->text('excerpt');
+                $table->text('body');
+                $table->boolean('featured')->default(false);
+                $table->string('image');
+                $table->timestamps();
+            });
+        }
+        if (!Schema::hasTable('category_post')) {
+            Schema::create('category_post', function(Blueprint $table) {
+                $table->id();
+                $table->foreignId('category_id')
+                    ->constrained()
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+                $table->foreignId('post_id')
+                    ->constrained()
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
+        if (!Schema::hasTable('post_tag')) {
+            Schema::create('post_tag', function(Blueprint $table) {
+                $table->id();
+                $table->foreignId('post_id')
+                    ->constrained()
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+                $table->foreignId('tag_id')
+                    ->constrained()
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
